Изменение : небольшое исправление для 2010 студии - оказывается она не может выполнять некоторые преобразования типов, в отличии от 2012 студии :)
git-svn-id: svn://mpc-next.ru/mpc-be/trunk@1613 d191f3aa-8897-45be-b9a4-56e5fb472b67

+++ src/ExtLib/Bento4/Core/Ap4TkhdAtom.h
@@ -84,7 +84,7 @@ public:
     AP4_UI32 GetHeight() const {return m_Height;}
 
     AP4_Integer GetRotation() {
-        AP4_Integer rotation = atan2((AP4_Integer)m_Matrix[1], (AP4_Integer)m_Matrix[0])*180.0/3.14159;
+        AP4_Integer rotation = atan2((double)(AP4_Integer)m_Matrix[1], (double)(AP4_Integer)m_Matrix[0])*180.0/3.14159;
         if (rotation < 0) {
 		    rotation += 360;
 		}
